fix runner warnings

+++ src/common/timeutil.h
@@ -51,7 +51,7 @@ namespace timeutil
 
         inline int64_t in_days(const std::time_t to, const std::time_t from)
         {
-            return static_cast<int64_t>(std::difftime(to, from) / (3600 * 24));
+            return static_cast<int64_t>(std::difftime(to, from) / (3600 * (int64_t)24));
         }
     }
 }
